+++ admin/module/komentar/index.php
@@ -14,7 +14,8 @@
       <div class="card-header">
 
         <i class="fas fa-table"></i>
-        Data Table Komentar</div>
+        Data Table Komentar
+      </div>
       <div class="card-body">
         <!-- <a href="index.php?page=module/komentar/tambah" class="btn btn-primary" style="float: right">Tambah Data</a> <br><br> -->
         <div class="table-responsive">
@@ -49,7 +50,7 @@
                   <td>
                     <!-- <a href="index.php?page=module/komentar/edit&id= <?php //echo $pecah->komentar_id 
                                                                           ?>" class="btn btn-warning">Edit</a> -->
-                    <a href="index.php?page=module/komentar/hapus&id= <?php echo $pecah->komentar_id ?>" class="btn btn-danger">Hapus</a>
+                    <a href="index.php?page=module/komentar/hapus&id= <?php echo $pecah->komentar_id ?>" class="btn btn-danger" onclick="confirm('Yakin Ingin Menghapus ?')">Hapus</a>
                   </td>
                 </tr>
               <?php } ?>
@@ -67,7 +68,7 @@
   <footer class="sticky-footer">
     <div class="container my-auto">
       <div class="copyright text-center my-auto">
-        <span>Copyright © Your Website 2019</span>
+        <span>Copyright © Kejaksaan Negeri <?= date("Y") ?></span>
       </div>
     </div>
   </footer>
